Feat/cocode (#59)
* feat(cocode-rs): add bash execution, context compaction, and error handling improvements

Shell Execution:
- Add comprehensive bash command execution with shell state management
- Implement shell snapshot/restore for environment preservation
- Support for command execution with proper PTY handling
- Add shell types and cleanup utilities

Context Compaction:
- Implement conversation compaction for long-running sessions
- Add compact config and compaction driver logic
- Support loop events for compaction progress tracking
- Enhance message history management

Error Handling Refactoring:
- Simplify error types across core/tools, core/api, core/context, core/prompt
- Streamline tool error handling with consistent patterns
- Fix error propagation in provider factory and retry logic

Subagent & Coordinator Improvements:
- Enhance subagent manager with improved lifecycle management
- Add coordinator manager enhancements for multi-agent execution
- Add tool context support for subagent spawning


* feat(core): add plan mode, system reminders, bash path extraction, and compaction

Plan Mode Support:
- Add plan-mode feature crate with state management, plan file ops, and slug generation
- Implement EnterPlanMode and ExitPlanMode tool handlers
- Add plan mode reminder generator with file read blocking integration

System Reminder Framework:
- Add system-reminder crate with orchestrator, generators, and throttling
- Implement generators: changed files, LSP diagnostics, nested memory, plan mode, todo reminders
- Add XML formatting, file tracking, and injection utilities

Bash Path Extraction:
- Add path extractor for shell command output analysis
- Implement readonly command detection and sandboxing improvements
- Add command parsing utilities

Context Compaction:
- Add compact_config module with strategy configuration
- Implement session memory agent for background extraction
- Add loop event tracking and conversation compaction logic
- Extend message history and tracking types

Tool Enhancements:
- Add edit/write tool blocking in plan mode
- Extend tool context with plan mode and reminder state


* feat(cocode): add TUI, hooks system, LSP, and model support

TUI Application (app/tui):
- Event-driven architecture with broker, handler, and stream components
- Session and UI state management with reactive updates
- Terminal abstraction with graceful cleanup
- Widgets: chat, input, status bar, tool panel
- Full keyboard navigation support

Hooks System (features/hooks):
- Async hook tracking with aggregator for concurrent execution
- Scope-based filtering for targeted hook activation
- Handler types: agent, command, prompt, webhook
- Registry for hook management and lifecycle
- Integration with tool executor and driver loop

LSP Integration (core/tools):
- LSP tool for language server interactions
- Context extensions for LSP manager access

Model Support (common/protocol, core/api):
- Role selection with role-based model configuration
- Thinking level conversion between unified and provider-specific formats
- Enhanced session state with model configuration

Infrastructure:
- arg0 crate for executable name detection
- CLI integration with TUI mode support
- System reminder generators for hook responses
- Cleanup of unused hyper-sdk options


* feat(cocode): add TUI, skills system, plugin architecture, and hyper-sdk enhancements

## TUI Application
- Add app.rs, render.rs, update.rs for TUI rendering and state management
- Add command.rs for TUI command handling
- Streamline CLAUDE.md documentation

## Skills System
- Add available_skills generator for system reminders
- Integrate skill execution in loop driver
- Enhance skill tool with context and execution tracking
- Update skill manager with expanded capabilities

## Plugin Architecture
- Add agent_loader, command_loader, mcp_loader for plugin loading
- Add registry.rs for plugin registration
- Add integration.rs for plugin integration
- Add command.rs for plugin commands
- Add mcp.rs for MCP protocol support
- Enhance contribution.rs with extended plugin contributions

## System Reminders
- Add generators: collab_notifications, delegate_mode, plan_mode_exit,
  plan_verification, token_usage, unified_tasks
- Expand generator.rs and config.rs
- Add types for reminder system

## Hyper SDK & Provider Enhancements
- Add model reasoning support with ReasoningEffort
- Enhance config resolver with expanded model resolution
- Add Gemini and OpenAI options for thinking/reasoning
- Improve thinking_convert for multi-provider support

## Subagent & Tools
- Add signal.rs for subagent signaling
- Enhance subagent manager
- Add result_persistence.rs for tool result storage
- Add tool_config.rs to protocol


* feat(cocode): add TUI runner, interceptors, justfile, and code refactoring

- Add TUI runner with ratatui integration for interactive sessions
- Implement HTTP interceptor chain for provider-sdks with byted-model-hub support
- Add comprehensive justfile with build, test, lint, and release tasks
- Refactor 200+ files: improve module re-exports and public API organization
- Add sandbox config helpers and session state management utilities
- Update rust-toolchain to 1.86


---------

+++ cocode-rs/app/cli/src/output.rs
@@ -3,7 +3,8 @@
 //! These functions provide formatted output for streaming responses,
 //! tool calls, errors, and session information.
 
-use std::io::{self, Write};
+use std::io::Write;
+use std::io::{self};
 
 /// Print a streaming text response.
 #[allow(dead_code)]
@@ -45,7 +46,7 @@ pub fn print_session_start(session_id: &str, model: &str, provider: &str) {
 }
 
 /// Print turn completion summary.
-pub fn print_turn_summary(input_tokens: i32, output_tokens: i32) {
+pub fn print_turn_summary(input_tokens: i64, output_tokens: i64) {
     println!();
     println!("[Tokens: {input_tokens} in / {output_tokens} out]");
 }
